reviewing orm

+++ likes/models.py
@@ -10,5 +10,5 @@ from django.contrib.contenttypes.fields import GenericForeignKey
 class LikedItem(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveBigIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
